requirements.txt
Django==2.2.11
django-crispy-forms==1.9.0
djangorestframework==3.11.0
pkg-resources==0.0.0
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.1
